refactor(child-pi): H-7 step 6 — extract spawn/env/args logic to child-pi-spawn.ts
Decomposed the child-process spawn pipeline into a dedicated module:
- BASE_ALLOWLIST: env var names always passed to children (now exported).
- buildChildPiSpawnOptions(): pure function that builds SpawnOptions from
  (cwd, env, model) — validates cwd, filters env via allowlist, validates
  NODE_PATH. Was already exported; re-exported from child-pi.ts to preserve
  the public API.
- assertOnlyControlEnvKeys(): runtime canary that verifies the merged env
  contains only PI_CREW_*/PI_TEAMS_* keys (defense against accidental
  secret leakage).
- prepareSpawnContext(): pre-spawn helper that builds worker args via
  buildPiWorkerArgs, attaches PI_CREW_STEERING_FILE, handles the pre-spawn
  abort check (returns immediate-abort result if parent signal already fired).
  Returns {spawnSpec, mergedEnv, tempDir} for the caller.

In runChildPi, replaced the inline spawn logic (~60 lines) with a single
prepareSpawnContext call. Removed now-unused imports (WINDOWS_ESSENTIAL_ENV_VARS,
buildScopedAllowList, sanitizeEnvSecrets, buildPiWorkerArgs, getPiSpawnCommand).

child-pi.ts: 1233 → 1105 lines (-128). New child-pi-spawn.ts: 273 lines.
All 83 child-pi + lock + dependent tests pass (no regressions).
